fix: clean up PR

+++ packages/builder/build-utils.ts
@@ -153,8 +153,11 @@ export async function setupReactNativeProject(
     .quiet();
 
 
-  console.log(`✓ Copying patches...`);
-  await $`cp -r ./patches ${appDir}`.cwd(__dirname);
+  // check if there are patches to copy
+  if (existsSync(join(__dirname, 'patches'))) {
+    console.log(`✓ Copying patches...`);
+    await $`cp -r ./patches ${appDir}`.cwd(__dirname);
+  }
 
   // Perform any library-specific setup before installing
   await installSetup(appDir, libraryName, 'preInstall');
@@ -187,8 +190,11 @@ export async function setupReactNativeProject(
   console.log('📦 Installing all dependencies...');
   await $`npm install`.cwd(appDir).quiet();
 
-  console.log('📦 Applying patches...');
-  await $`npx patch-package`.cwd(appDir);
+  // check if there is patches folder to apply
+  if (existsSync(join(appDir, 'patches'))) {
+    console.log('📦 Applying patches...');
+    await $`npx patch-package`.cwd(appDir);
+  }
 
   // Check if the react-native version is correctly set
   checkRnVersion(appDir, reactNativeVersion);
